Joint Selection Rate

Calculate the Joint Selection Rate (JSR) for polytomous data. JSR measures the proportion of respondents who selected specific category combinations between pairs of items. For each pair of items (j,k), it returns a contingency table showing the joint probability of selecting each category combination.

Usage

JSR(U, na = NULL, Z = NULL, w = NULL)

Arguments

U

Either an object of class "exametrika" or raw data. When raw data is given, it is converted to the exametrika class with the dataFormat function.

na

Values to be treated as missing values.

Z

Missing indicator matrix of type matrix or data.frame. Values of 1 indicate observed responses, while 0 indicates missing data.

w

Item weight vector specifying the relative importance of each item.

Value

A list of Joint Selection Rate matrices for each item pair.

Examples

# example code
# Calculate JCRR using sample dataset J5S1000
JSR(J5S1000)
#> [[1]]
#> [[1]][[1]]
#>          
#>             V1-Cat1   V1-Cat2   V1-Cat3   V1-Cat4
#>   V1-Cat1 0.2104733 0.0000000 0.0000000 0.0000000
#>   V1-Cat2 0.0000000 0.3011078 0.0000000 0.0000000
#>   V1-Cat3 0.0000000 0.0000000 0.2779456 0.0000000
#>   V1-Cat4 0.0000000 0.0000000 0.0000000 0.2104733
#> 
#> [[1]][[2]]
#>          
#>              V2-Cat1    V2-Cat2    V2-Cat3
#>   V1-Cat1 0.09053917 0.08443540 0.03560529
#>   V1-Cat2 0.07934893 0.16073245 0.06103764
#>   V1-Cat3 0.07222787 0.13326551 0.07426246
#>   V1-Cat4 0.03865717 0.08952187 0.08036623
#> 
#> [[1]][[3]]
#>          
#>              V3-Cat1    V3-Cat2    V3-Cat3    V3-Cat4
#>   V1-Cat1 0.06774520 0.07077856 0.04347826 0.02932255
#>   V1-Cat2 0.06167846 0.08392315 0.07886754 0.07785642
#>   V1-Cat3 0.04651163 0.07684530 0.08695652 0.06572295
#>   V1-Cat4 0.02628918 0.04954499 0.05561173 0.07886754
#> 
#> [[1]][[4]]
#>          
#>              V4-Cat1    V4-Cat2    V4-Cat3
#>   V1-Cat1 0.08787879 0.08585859 0.03636364
#>   V1-Cat2 0.09090909 0.13333333 0.07777778
#>   V1-Cat3 0.05656566 0.14242424 0.07979798
#>   V1-Cat4 0.02626263 0.09494949 0.08787879
#> 
#> [[1]][[5]]
#>          
#>              V5-Cat1    V5-Cat2    V5-Cat3    V5-Cat4
#>   V1-Cat1 0.07366297 0.06458123 0.04641776 0.02623613
#>   V1-Cat2 0.08173562 0.08980827 0.07467205 0.05549950
#>   V1-Cat3 0.05751766 0.06861756 0.09081736 0.06054490
#>   V1-Cat4 0.02421796 0.04540868 0.06962664 0.07063572
#> 
#> 
#> [[2]]
#> [[2]][[1]]
#>          
#>              V1-Cat1    V1-Cat2    V1-Cat3    V1-Cat4
#>   V2-Cat1 0.09053917 0.07934893 0.07222787 0.03865717
#>   V2-Cat2 0.08443540 0.16073245 0.13326551 0.08952187
#>   V2-Cat3 0.03560529 0.06103764 0.07426246 0.08036623
#> 
#> [[2]][[2]]
#>          
#>             V2-Cat1   V2-Cat2   V2-Cat3
#>   V2-Cat1 0.2808081 0.0000000 0.0000000
#>   V2-Cat2 0.0000000 0.4676768 0.0000000
#>   V2-Cat3 0.0000000 0.0000000 0.2515152
#> 
#> [[2]][[3]]
#>          
#>              V3-Cat1    V3-Cat2    V3-Cat3    V3-Cat4
#>   V2-Cat1 0.08113590 0.09736308 0.05882353 0.04462475
#>   V2-Cat2 0.08316430 0.13590264 0.13184584 0.11561866
#>   V2-Cat3 0.03752535 0.04868154 0.07302231 0.09229209
#> 
#> [[2]][[4]]
#>          
#>              V4-Cat1    V4-Cat2    V4-Cat3
#>   V2-Cat1 0.11955420 0.12867275 0.03242148
#>   V2-Cat2 0.11854103 0.21479230 0.13373860
#>   V2-Cat3 0.02228977 0.11651469 0.11347518
#> 
#> [[2]][[5]]
#>          
#>              V5-Cat1    V5-Cat2    V5-Cat3    V5-Cat4
#>   V2-Cat1 0.09311741 0.08097166 0.06983806 0.03744939
#>   V2-Cat2 0.11032389 0.14170040 0.12449393 0.09109312
#>   V2-Cat3 0.03238866 0.04655870 0.08704453 0.08502024
#> 
#> 
#> [[3]]
#> [[3]][[1]]
#>          
#>              V1-Cat1    V1-Cat2    V1-Cat3    V1-Cat4
#>   V3-Cat1 0.06774520 0.06167846 0.04651163 0.02628918
#>   V3-Cat2 0.07077856 0.08392315 0.07684530 0.04954499
#>   V3-Cat3 0.04347826 0.07886754 0.08695652 0.05561173
#>   V3-Cat4 0.02932255 0.07785642 0.06572295 0.07886754
#> 
#> [[3]][[2]]
#>          
#>              V2-Cat1    V2-Cat2    V2-Cat3
#>   V3-Cat1 0.08113590 0.08316430 0.03752535
#>   V3-Cat2 0.09736308 0.13590264 0.04868154
#>   V3-Cat3 0.05882353 0.13184584 0.07302231
#>   V3-Cat4 0.04462475 0.11561866 0.09229209
#> 
#> [[3]][[3]]
#>          
#>             V3-Cat1   V3-Cat2   V3-Cat3   V3-Cat4
#>   V3-Cat1 0.2028112 0.0000000 0.0000000 0.0000000
#>   V3-Cat2 0.0000000 0.2811245 0.0000000 0.0000000
#>   V3-Cat3 0.0000000 0.0000000 0.2630522 0.0000000
#>   V3-Cat4 0.0000000 0.0000000 0.0000000 0.2530120
#> 
#> [[3]][[4]]
#>          
#>              V4-Cat1    V4-Cat2    V4-Cat3
#>   V3-Cat1 0.09063444 0.08056395 0.03222558
#>   V3-Cat2 0.08660624 0.13997986 0.05438066
#>   V3-Cat3 0.04431017 0.12487412 0.09264854
#>   V3-Cat4 0.04028197 0.11178248 0.10171198
#> 
#> [[3]][[5]]
#>          
#>              V5-Cat1    V5-Cat2    V5-Cat3    V5-Cat4
#>   V3-Cat1 0.06941650 0.05331992 0.04728370 0.03319920
#>   V3-Cat2 0.08249497 0.07645875 0.07847082 0.04426559
#>   V3-Cat3 0.05432596 0.07042254 0.06639839 0.07142857
#>   V3-Cat4 0.03219316 0.06740443 0.08953722 0.06338028
#> 
#> 
#> [[4]]
#> [[4]][[1]]
#>          
#>              V1-Cat1    V1-Cat2    V1-Cat3    V1-Cat4
#>   V4-Cat1 0.08787879 0.09090909 0.05656566 0.02626263
#>   V4-Cat2 0.08585859 0.13333333 0.14242424 0.09494949
#>   V4-Cat3 0.03636364 0.07777778 0.07979798 0.08787879
#> 
#> [[4]][[2]]
#>          
#>              V2-Cat1    V2-Cat2    V2-Cat3
#>   V4-Cat1 0.11955420 0.11854103 0.02228977
#>   V4-Cat2 0.12867275 0.21479230 0.11651469
#>   V4-Cat3 0.03242148 0.13373860 0.11347518
#> 
#> [[4]][[3]]
#>          
#>              V3-Cat1    V3-Cat2    V3-Cat3    V3-Cat4
#>   V4-Cat1 0.09063444 0.08660624 0.04431017 0.04028197
#>   V4-Cat2 0.08056395 0.13997986 0.12487412 0.11178248
#>   V4-Cat3 0.03222558 0.05438066 0.09264854 0.10171198
#> 
#> [[4]][[4]]
#>          
#>             V4-Cat1   V4-Cat2   V4-Cat3
#>   V4-Cat1 0.2617854 0.0000000 0.0000000
#>   V4-Cat2 0.0000000 0.4573721 0.0000000
#>   V4-Cat3 0.0000000 0.0000000 0.2808425
#> 
#> [[4]][[5]]
#>          
#>              V5-Cat1    V5-Cat2    V5-Cat3    V5-Cat4
#>   V4-Cat1 0.09447236 0.07437186 0.06231156 0.03115578
#>   V4-Cat2 0.10351759 0.13467337 0.12663317 0.09246231
#>   V4-Cat3 0.03919598 0.05728643 0.09346734 0.09045226
#> 
#> 
#> [[5]]
#> [[5]][[1]]
#>          
#>              V1-Cat1    V1-Cat2    V1-Cat3    V1-Cat4
#>   V5-Cat1 0.07366297 0.08173562 0.05751766 0.02421796
#>   V5-Cat2 0.06458123 0.08980827 0.06861756 0.04540868
#>   V5-Cat3 0.04641776 0.07467205 0.09081736 0.06962664
#>   V5-Cat4 0.02623613 0.05549950 0.06054490 0.07063572
#> 
#> [[5]][[2]]
#>          
#>              V2-Cat1    V2-Cat2    V2-Cat3
#>   V5-Cat1 0.09311741 0.11032389 0.03238866
#>   V5-Cat2 0.08097166 0.14170040 0.04655870
#>   V5-Cat3 0.06983806 0.12449393 0.08704453
#>   V5-Cat4 0.03744939 0.09109312 0.08502024
#> 
#> [[5]][[3]]
#>          
#>              V3-Cat1    V3-Cat2    V3-Cat3    V3-Cat4
#>   V5-Cat1 0.06941650 0.08249497 0.05432596 0.03219316
#>   V5-Cat2 0.05331992 0.07645875 0.07042254 0.06740443
#>   V5-Cat3 0.04728370 0.07847082 0.06639839 0.08953722
#>   V5-Cat4 0.03319920 0.04426559 0.07142857 0.06338028
#> 
#> [[5]][[4]]
#>          
#>              V4-Cat1    V4-Cat2    V4-Cat3
#>   V5-Cat1 0.09447236 0.10351759 0.03919598
#>   V5-Cat2 0.07437186 0.13467337 0.05728643
#>   V5-Cat3 0.06231156 0.12663317 0.09346734
#>   V5-Cat4 0.03115578 0.09246231 0.09045226
#> 
#> [[5]][[5]]
#>          
#>             V5-Cat1   V5-Cat2   V5-Cat3   V5-Cat4
#>   V5-Cat1 0.2374749 0.0000000 0.0000000 0.0000000
#>   V5-Cat2 0.0000000 0.2675351 0.0000000 0.0000000
#>   V5-Cat3 0.0000000 0.0000000 0.2815631 0.0000000
#>   V5-Cat4 0.0000000 0.0000000 0.0000000 0.2134269
#> 
#>