Number Right Score

The Number-Right Score (NRS) function calculates the weighted sum of correct responses for each examinee. This function is applicable only to binary response data.

For each examinee, the score is computed as: $$NRS_i = \sum_{j=1}^J z_{ij}u_{ij}w_j$$ where:

\(z_{ij}\) is the missing response indicator (0/1)

\(u_{ij}\) is the response (0/1)

\(w_j\) is the item weight

Usage

nrs(U, na = NULL, Z = NULL, w = NULL)

# Default S3 method
nrs(U, na = NULL, Z = NULL, w = NULL)

# S3 method for class 'binary'
nrs(U, na = NULL, Z = NULL, w = NULL)

Arguments

U

Either an object of class "exametrika" or raw data. When raw data is given, it is converted to the exametrika class with the dataFormat function.

na

Values to be treated as missing values.

Z

Missing indicator matrix of type matrix or data.frame. Values of 1 indicate observed responses, while 0 indicates missing data.

w

Item weight vector specifying the relative importance of each item.

Value

A numeric vector containing the Number-Right Score for each examinee. The score represents the weighted sum of correct answers, where:

Maximum score is the sum of all item weights

Minimum score is 0

Missing responses do not contribute to the score

Note

This function is implemented using a binary data compatibility wrapper and will raise an error if used with polytomous data.

Examples

# using sample dataset
nrs(J15S500)
#>        [,1]
#>   [1,]    8
#>   [2,]   10
#>   [3,]   10
#>   [4,]   11
#>   [5,]    6
#>   [6,]   12
#>   [7,]    9
#>   [8,]   10
#>   [9,]   14
#>  [10,]    9
#>  [11,]    9
#>  [12,]    8
#>  [13,]   13
#>  [14,]   12
#>  [15,]    5
#>  [16,]    5
#>  [17,]   12
#>  [18,]    7
#>  [19,]   12
#>  [20,]    6
#>  [21,]   13
#>  [22,]    8
#>  [23,]   10
#>  [24,]   10
#>  [25,]   10
#>  [26,]    5
#>  [27,]    9
#>  [28,]   11
#>  [29,]   11
#>  [30,]    4
#>  [31,]    5
#>  [32,]   11
#>  [33,]    5
#>  [34,]    5
#>  [35,]   15
#>  [36,]    8
#>  [37,]   12
#>  [38,]   11
#>  [39,]    8
#>  [40,]   10
#>  [41,]   11
#>  [42,]   14
#>  [43,]   10
#>  [44,]    9
#>  [45,]   12
#>  [46,]   11
#>  [47,]   10
#>  [48,]   12
#>  [49,]   11
#>  [50,]   13
#>  [51,]   11
#>  [52,]   12
#>  [53,]   10
#>  [54,]   11
#>  [55,]   13
#>  [56,]   12
#>  [57,]    9
#>  [58,]   11
#>  [59,]   14
#>  [60,]   12
#>  [61,]   14
#>  [62,]   11
#>  [63,]    9
#>  [64,]   11
#>  [65,]   10
#>  [66,]    5
#>  [67,]    7
#>  [68,]    6
#>  [69,]   12
#>  [70,]    6
#>  [71,]    5
#>  [72,]   10
#>  [73,]   10
#>  [74,]   10
#>  [75,]   12
#>  [76,]    8
#>  [77,]   13
#>  [78,]   12
#>  [79,]   12
#>  [80,]    4
#>  [81,]    9
#>  [82,]    5
#>  [83,]    8
#>  [84,]    8
#>  [85,]   12
#>  [86,]    8
#>  [87,]   11
#>  [88,]    8
#>  [89,]   12
#>  [90,]   11
#>  [91,]    7
#>  [92,]    6
#>  [93,]    8
#>  [94,]   12
#>  [95,]    8
#>  [96,]   11
#>  [97,]   10
#>  [98,]    6
#>  [99,]   11
#> [100,]   12
#> [101,]   10
#> [102,]    6
#> [103,]   13
#> [104,]    5
#> [105,]   11
#> [106,]    7
#> [107,]   11
#> [108,]   11
#> [109,]    6
#> [110,]   12
#> [111,]   12
#> [112,]   11
#> [113,]    6
#> [114,]    8
#> [115,]    5
#> [116,]   13
#> [117,]   11
#> [118,]    7
#> [119,]    8
#> [120,]    8
#> [121,]   10
#> [122,]    8
#> [123,]   12
#> [124,]   13
#> [125,]    9
#> [126,]   13
#> [127,]    4
#> [128,]   11
#> [129,]   12
#> [130,]   12
#> [131,]   10
#> [132,]   10
#> [133,]   11
#> [134,]    5
#> [135,]    8
#> [136,]    5
#> [137,]   13
#> [138,]   11
#> [139,]   11
#> [140,]   10
#> [141,]   11
#> [142,]    5
#> [143,]   13
#> [144,]    4
#> [145,]   12
#> [146,]    7
#> [147,]    5
#> [148,]   12
#> [149,]    9
#> [150,]   13
#> [151,]   13
#> [152,]   11
#> [153,]   14
#> [154,]    9
#> [155,]    6
#> [156,]   11
#> [157,]   10
#> [158,]   13
#> [159,]   13
#> [160,]   15
#> [161,]   11
#> [162,]    7
#> [163,]    7
#> [164,]    5
#> [165,]   10
#> [166,]    5
#> [167,]   10
#> [168,]    8
#> [169,]   14
#> [170,]    6
#> [171,]    6
#> [172,]   13
#> [173,]   10
#> [174,]   12
#> [175,]    7
#> [176,]   10
#> [177,]   11
#> [178,]    8
#> [179,]   10
#> [180,]    8
#> [181,]    7
#> [182,]    6
#> [183,]   11
#> [184,]    8
#> [185,]   11
#> [186,]   11
#> [187,]    8
#> [188,]   11
#> [189,]    9
#> [190,]   10
#> [191,]   10
#> [192,]    7
#> [193,]   10
#> [194,]    6
#> [195,]    3
#> [196,]   12
#> [197,]   11
#> [198,]   13
#> [199,]   10
#> [200,]   11
#> [201,]   12
#> [202,]   11
#> [203,]    5
#> [204,]   12
#> [205,]   10
#> [206,]    8
#> [207,]   12
#> [208,]   11
#> [209,]    9
#> [210,]   11
#> [211,]   10
#> [212,]   14
#> [213,]   10
#> [214,]    8
#> [215,]   10
#> [216,]    9
#> [217,]   13
#> [218,]   12
#> [219,]    4
#> [220,]    6
#> [221,]    5
#> [222,]   11
#> [223,]    9
#> [224,]    9
#> [225,]   10
#> [226,]   10
#> [227,]    9
#> [228,]   11
#> [229,]    7
#> [230,]   10
#> [231,]    7
#> [232,]    4
#> [233,]   12
#> [234,]    4
#> [235,]    9
#> [236,]    9
#> [237,]    7
#> [238,]   14
#> [239,]   12
#> [240,]   11
#> [241,]   12
#> [242,]   11
#> [243,]    9
#> [244,]    8
#> [245,]   11
#> [246,]    5
#> [247,]   11
#> [248,]   15
#> [249,]    8
#> [250,]   14
#> [251,]    9
#> [252,]    7
#> [253,]   11
#> [254,]   10
#> [255,]   11
#> [256,]   10
#> [257,]    8
#> [258,]    7
#> [259,]    7
#> [260,]    9
#> [261,]    7
#> [262,]   13
#> [263,]   12
#> [264,]    8
#> [265,]   12
#> [266,]   11
#> [267,]   11
#> [268,]    8
#> [269,]   12
#> [270,]    8
#> [271,]   12
#> [272,]    9
#> [273,]    6
#> [274,]    9
#> [275,]   10
#> [276,]   11
#> [277,]    7
#> [278,]    6
#> [279,]   10
#> [280,]    7
#> [281,]   10
#> [282,]   12
#> [283,]    7
#> [284,]   12
#> [285,]    9
#> [286,]    6
#> [287,]   13
#> [288,]    9
#> [289,]   13
#> [290,]   15
#> [291,]   10
#> [292,]   14
#> [293,]    4
#> [294,]   12
#> [295,]    6
#> [296,]    7
#> [297,]   11
#> [298,]   12
#> [299,]    5
#> [300,]   11
#> [301,]    7
#> [302,]   13
#> [303,]    9
#> [304,]    9
#> [305,]   10
#> [306,]   12
#> [307,]    6
#> [308,]   12
#> [309,]   11
#> [310,]    9
#> [311,]   10
#> [312,]   11
#> [313,]   12
#> [314,]    7
#> [315,]   11
#> [316,]   11
#> [317,]   13
#> [318,]    7
#> [319,]    6
#> [320,]    9
#> [321,]    9
#> [322,]   13
#> [323,]    9
#> [324,]   13
#> [325,]   13
#> [326,]   13
#> [327,]    8
#> [328,]   10
#> [329,]   11
#> [330,]   10
#> [331,]   12
#> [332,]    7
#> [333,]    9
#> [334,]   13
#> [335,]   11
#> [336,]   10
#> [337,]   10
#> [338,]   12
#> [339,]   13
#> [340,]    9
#> [341,]    4
#> [342,]   10
#> [343,]   10
#> [344,]   14
#> [345,]    7
#> [346,]   11
#> [347,]   10
#> [348,]   12
#> [349,]   10
#> [350,]   12
#> [351,]   10
#> [352,]    8
#> [353,]   11
#> [354,]    8
#> [355,]   12
#> [356,]   13
#> [357,]    7
#> [358,]   14
#> [359,]    2
#> [360,]    9
#> [361,]   12
#> [362,]   13
#> [363,]    7
#> [364,]   11
#> [365,]   10
#> [366,]   11
#> [367,]   15
#> [368,]    6
#> [369,]   10
#> [370,]   10
#> [371,]   14
#> [372,]   11
#> [373,]   13
#> [374,]    9
#> [375,]    9
#> [376,]   10
#> [377,]   10
#> [378,]   10
#> [379,]   14
#> [380,]   12
#> [381,]    2
#> [382,]    7
#> [383,]    7
#> [384,]    7
#> [385,]   12
#> [386,]    8
#> [387,]    8
#> [388,]    7
#> [389,]   12
#> [390,]    9
#> [391,]    6
#> [392,]    7
#> [393,]   11
#> [394,]    7
#> [395,]   11
#> [396,]    8
#> [397,]   11
#> [398,]   12
#> [399,]    6
#> [400,]    5
#> [401,]   13
#> [402,]    7
#> [403,]   10
#> [404,]   12
#> [405,]   10
#> [406,]   10
#> [407,]    2
#> [408,]   11
#> [409,]   10
#> [410,]   12
#> [411,]   12
#> [412,]   10
#> [413,]   10
#> [414,]    9
#> [415,]   10
#> [416,]   15
#> [417,]    9
#> [418,]   13
#> [419,]    6
#> [420,]    9
#> [421,]    9
#> [422,]    7
#> [423,]    9
#> [424,]   13
#> [425,]    9
#> [426,]    9
#> [427,]   13
#> [428,]    5
#> [429,]   11
#> [430,]   11
#> [431,]   10
#> [432,]   12
#> [433,]   11
#> [434,]   13
#> [435,]   10
#> [436,]   14
#> [437,]   11
#> [438,]   10
#> [439,]    8
#> [440,]   11
#> [441,]   11
#> [442,]   10
#> [443,]   11
#> [444,]   11
#> [445,]    8
#> [446,]    6
#> [447,]   14
#> [448,]   12
#> [449,]   10
#> [450,]   13
#> [451,]   10
#> [452,]   13
#> [453,]    8
#> [454,]   12
#> [455,]   13
#> [456,]    4
#> [457,]   13
#> [458,]   11
#> [459,]   10
#> [460,]    6
#> [461,]    7
#> [462,]    8
#> [463,]    6
#> [464,]    3
#> [465,]   10
#> [466,]   13
#> [467,]   12
#> [468,]   13
#> [469,]    5
#> [470,]    6
#> [471,]    5
#> [472,]    6
#> [473,]   10
#> [474,]    6
#> [475,]   12
#> [476,]    9
#> [477,]   11
#> [478,]   14
#> [479,]    7
#> [480,]   12
#> [481,]    7
#> [482,]   13
#> [483,]    9
#> [484,]   11
#> [485,]   14
#> [486,]   13
#> [487,]   11
#> [488,]   11
#> [489,]   11
#> [490,]    9
#> [491,]    6
#> [492,]   12
#> [493,]    9
#> [494,]    9
#> [495,]    7
#> [496,]    6
#> [497,]    7
#> [498,]    7
#> [499,]   10
#> [500,]    7