Passage Rate of Student

The Passage Rate for each student is calculated as their Number-Right Score (NRS) divided by the number of items presented to them. This function is applicable only to binary response data.

The passage rate is calculated as: $$P_i = \frac{\sum_{j=1}^J z_{ij}u_{ij}w_j}{\sum_{j=1}^J z_{ij}}$$ where:

\(z_{ij}\) is the missing response indicator (0/1)

\(u_{ij}\) is the response (0/1)

\(w_j\) is the item weight

Usage

passage(U, na = NULL, Z = NULL, w = NULL)

# Default S3 method
passage(U, na = NULL, Z = NULL, w = NULL)

# S3 method for class 'binary'
passage(U, na = NULL, Z = NULL, w = NULL)

Arguments

U

Either an object of class "exametrika" or raw data. When raw data is given, it is converted to the exametrika class with the dataFormat function.

na

Values to be treated as missing values.

Z

Missing indicator matrix of type matrix or data.frame. Values of 1 indicate observed responses, while 0 indicates missing data.

w

Item weight vector specifying the relative importance of each item.

Value

A numeric vector containing the passage rate for each student. Values range from 0 to 1 (or maximum weight) where:

1: Perfect score on all attempted items

0: No correct answers

NA: No items attempted

Note

This function is implemented using a binary data compatibility wrapper and will raise an error if used with polytomous data.

The passage rate accounts for missing responses by only considering items that were actually presented to each student. This provides a fair comparison between students who attempted different numbers of items.

Examples

# using sample dataset
passage(J15S500)
#>             [,1]
#>   [1,] 0.5333333
#>   [2,] 0.6666667
#>   [3,] 0.6666667
#>   [4,] 0.7333333
#>   [5,] 0.4000000
#>   [6,] 0.8000000
#>   [7,] 0.6000000
#>   [8,] 0.6666667
#>   [9,] 0.9333333
#>  [10,] 0.6000000
#>  [11,] 0.6000000
#>  [12,] 0.5333333
#>  [13,] 0.8666667
#>  [14,] 0.8000000
#>  [15,] 0.3333333
#>  [16,] 0.3333333
#>  [17,] 0.8000000
#>  [18,] 0.4666667
#>  [19,] 0.8000000
#>  [20,] 0.4000000
#>  [21,] 0.8666667
#>  [22,] 0.5333333
#>  [23,] 0.6666667
#>  [24,] 0.6666667
#>  [25,] 0.6666667
#>  [26,] 0.3333333
#>  [27,] 0.6000000
#>  [28,] 0.7333333
#>  [29,] 0.7333333
#>  [30,] 0.2666667
#>  [31,] 0.3333333
#>  [32,] 0.7333333
#>  [33,] 0.3333333
#>  [34,] 0.3333333
#>  [35,] 1.0000000
#>  [36,] 0.5333333
#>  [37,] 0.8000000
#>  [38,] 0.7333333
#>  [39,] 0.5333333
#>  [40,] 0.6666667
#>  [41,] 0.7333333
#>  [42,] 0.9333333
#>  [43,] 0.6666667
#>  [44,] 0.6000000
#>  [45,] 0.8000000
#>  [46,] 0.7333333
#>  [47,] 0.6666667
#>  [48,] 0.8000000
#>  [49,] 0.7333333
#>  [50,] 0.8666667
#>  [51,] 0.7333333
#>  [52,] 0.8000000
#>  [53,] 0.6666667
#>  [54,] 0.7333333
#>  [55,] 0.8666667
#>  [56,] 0.8000000
#>  [57,] 0.6000000
#>  [58,] 0.7333333
#>  [59,] 0.9333333
#>  [60,] 0.8000000
#>  [61,] 0.9333333
#>  [62,] 0.7333333
#>  [63,] 0.6000000
#>  [64,] 0.7333333
#>  [65,] 0.6666667
#>  [66,] 0.3333333
#>  [67,] 0.4666667
#>  [68,] 0.4000000
#>  [69,] 0.8000000
#>  [70,] 0.4000000
#>  [71,] 0.3333333
#>  [72,] 0.6666667
#>  [73,] 0.6666667
#>  [74,] 0.6666667
#>  [75,] 0.8000000
#>  [76,] 0.5333333
#>  [77,] 0.8666667
#>  [78,] 0.8000000
#>  [79,] 0.8000000
#>  [80,] 0.2666667
#>  [81,] 0.6000000
#>  [82,] 0.3333333
#>  [83,] 0.5333333
#>  [84,] 0.5333333
#>  [85,] 0.8000000
#>  [86,] 0.5333333
#>  [87,] 0.7333333
#>  [88,] 0.5333333
#>  [89,] 0.8000000
#>  [90,] 0.7333333
#>  [91,] 0.4666667
#>  [92,] 0.4000000
#>  [93,] 0.5333333
#>  [94,] 0.8000000
#>  [95,] 0.5333333
#>  [96,] 0.7333333
#>  [97,] 0.6666667
#>  [98,] 0.4000000
#>  [99,] 0.7333333
#> [100,] 0.8000000
#> [101,] 0.6666667
#> [102,] 0.4000000
#> [103,] 0.8666667
#> [104,] 0.3333333
#> [105,] 0.7333333
#> [106,] 0.4666667
#> [107,] 0.7333333
#> [108,] 0.7333333
#> [109,] 0.4000000
#> [110,] 0.8000000
#> [111,] 0.8000000
#> [112,] 0.7333333
#> [113,] 0.4000000
#> [114,] 0.5333333
#> [115,] 0.3333333
#> [116,] 0.8666667
#> [117,] 0.7333333
#> [118,] 0.4666667
#> [119,] 0.5333333
#> [120,] 0.5333333
#> [121,] 0.6666667
#> [122,] 0.5333333
#> [123,] 0.8000000
#> [124,] 0.8666667
#> [125,] 0.6000000
#> [126,] 0.8666667
#> [127,] 0.2666667
#> [128,] 0.7333333
#> [129,] 0.8000000
#> [130,] 0.8000000
#> [131,] 0.6666667
#> [132,] 0.6666667
#> [133,] 0.7333333
#> [134,] 0.3333333
#> [135,] 0.5333333
#> [136,] 0.3333333
#> [137,] 0.8666667
#> [138,] 0.7333333
#> [139,] 0.7333333
#> [140,] 0.6666667
#> [141,] 0.7333333
#> [142,] 0.3333333
#> [143,] 0.8666667
#> [144,] 0.2666667
#> [145,] 0.8000000
#> [146,] 0.4666667
#> [147,] 0.3333333
#> [148,] 0.8000000
#> [149,] 0.6000000
#> [150,] 0.8666667
#> [151,] 0.8666667
#> [152,] 0.7333333
#> [153,] 0.9333333
#> [154,] 0.6000000
#> [155,] 0.4000000
#> [156,] 0.7333333
#> [157,] 0.6666667
#> [158,] 0.8666667
#> [159,] 0.8666667
#> [160,] 1.0000000
#> [161,] 0.7333333
#> [162,] 0.4666667
#> [163,] 0.4666667
#> [164,] 0.3333333
#> [165,] 0.6666667
#> [166,] 0.3333333
#> [167,] 0.6666667
#> [168,] 0.5333333
#> [169,] 0.9333333
#> [170,] 0.4000000
#> [171,] 0.4000000
#> [172,] 0.8666667
#> [173,] 0.6666667
#> [174,] 0.8000000
#> [175,] 0.4666667
#> [176,] 0.6666667
#> [177,] 0.7333333
#> [178,] 0.5333333
#> [179,] 0.6666667
#> [180,] 0.5333333
#> [181,] 0.4666667
#> [182,] 0.4000000
#> [183,] 0.7333333
#> [184,] 0.5333333
#> [185,] 0.7333333
#> [186,] 0.7333333
#> [187,] 0.5333333
#> [188,] 0.7333333
#> [189,] 0.6000000
#> [190,] 0.6666667
#> [191,] 0.6666667
#> [192,] 0.4666667
#> [193,] 0.6666667
#> [194,] 0.4000000
#> [195,] 0.2000000
#> [196,] 0.8000000
#> [197,] 0.7333333
#> [198,] 0.8666667
#> [199,] 0.6666667
#> [200,] 0.7333333
#> [201,] 0.8000000
#> [202,] 0.7333333
#> [203,] 0.3333333
#> [204,] 0.8000000
#> [205,] 0.6666667
#> [206,] 0.5333333
#> [207,] 0.8000000
#> [208,] 0.7333333
#> [209,] 0.6000000
#> [210,] 0.7333333
#> [211,] 0.6666667
#> [212,] 0.9333333
#> [213,] 0.6666667
#> [214,] 0.5333333
#> [215,] 0.6666667
#> [216,] 0.6000000
#> [217,] 0.8666667
#> [218,] 0.8000000
#> [219,] 0.2666667
#> [220,] 0.4000000
#> [221,] 0.3333333
#> [222,] 0.7333333
#> [223,] 0.6000000
#> [224,] 0.6000000
#> [225,] 0.6666667
#> [226,] 0.6666667
#> [227,] 0.6000000
#> [228,] 0.7333333
#> [229,] 0.4666667
#> [230,] 0.6666667
#> [231,] 0.4666667
#> [232,] 0.2666667
#> [233,] 0.8000000
#> [234,] 0.2666667
#> [235,] 0.6000000
#> [236,] 0.6000000
#> [237,] 0.4666667
#> [238,] 0.9333333
#> [239,] 0.8000000
#> [240,] 0.7333333
#> [241,] 0.8000000
#> [242,] 0.7333333
#> [243,] 0.6000000
#> [244,] 0.5333333
#> [245,] 0.7333333
#> [246,] 0.3333333
#> [247,] 0.7333333
#> [248,] 1.0000000
#> [249,] 0.5333333
#> [250,] 0.9333333
#> [251,] 0.6000000
#> [252,] 0.4666667
#> [253,] 0.7333333
#> [254,] 0.6666667
#> [255,] 0.7333333
#> [256,] 0.6666667
#> [257,] 0.5333333
#> [258,] 0.4666667
#> [259,] 0.4666667
#> [260,] 0.6000000
#> [261,] 0.4666667
#> [262,] 0.8666667
#> [263,] 0.8000000
#> [264,] 0.5333333
#> [265,] 0.8000000
#> [266,] 0.7333333
#> [267,] 0.7333333
#> [268,] 0.5333333
#> [269,] 0.8000000
#> [270,] 0.5333333
#> [271,] 0.8000000
#> [272,] 0.6000000
#> [273,] 0.4000000
#> [274,] 0.6000000
#> [275,] 0.6666667
#> [276,] 0.7333333
#> [277,] 0.4666667
#> [278,] 0.4000000
#> [279,] 0.6666667
#> [280,] 0.4666667
#> [281,] 0.6666667
#> [282,] 0.8000000
#> [283,] 0.4666667
#> [284,] 0.8000000
#> [285,] 0.6000000
#> [286,] 0.4000000
#> [287,] 0.8666667
#> [288,] 0.6000000
#> [289,] 0.8666667
#> [290,] 1.0000000
#> [291,] 0.6666667
#> [292,] 0.9333333
#> [293,] 0.2666667
#> [294,] 0.8000000
#> [295,] 0.4000000
#> [296,] 0.4666667
#> [297,] 0.7333333
#> [298,] 0.8000000
#> [299,] 0.3333333
#> [300,] 0.7333333
#> [301,] 0.4666667
#> [302,] 0.8666667
#> [303,] 0.6000000
#> [304,] 0.6000000
#> [305,] 0.6666667
#> [306,] 0.8000000
#> [307,] 0.4000000
#> [308,] 0.8000000
#> [309,] 0.7333333
#> [310,] 0.6000000
#> [311,] 0.6666667
#> [312,] 0.7333333
#> [313,] 0.8000000
#> [314,] 0.4666667
#> [315,] 0.7333333
#> [316,] 0.7333333
#> [317,] 0.8666667
#> [318,] 0.4666667
#> [319,] 0.4000000
#> [320,] 0.6000000
#> [321,] 0.6000000
#> [322,] 0.8666667
#> [323,] 0.6000000
#> [324,] 0.8666667
#> [325,] 0.8666667
#> [326,] 0.8666667
#> [327,] 0.5333333
#> [328,] 0.6666667
#> [329,] 0.7333333
#> [330,] 0.6666667
#> [331,] 0.8000000
#> [332,] 0.4666667
#> [333,] 0.6000000
#> [334,] 0.8666667
#> [335,] 0.7333333
#> [336,] 0.6666667
#> [337,] 0.6666667
#> [338,] 0.8000000
#> [339,] 0.8666667
#> [340,] 0.6000000
#> [341,] 0.2666667
#> [342,] 0.6666667
#> [343,] 0.6666667
#> [344,] 0.9333333
#> [345,] 0.4666667
#> [346,] 0.7333333
#> [347,] 0.6666667
#> [348,] 0.8000000
#> [349,] 0.6666667
#> [350,] 0.8000000
#> [351,] 0.6666667
#> [352,] 0.5333333
#> [353,] 0.7333333
#> [354,] 0.5333333
#> [355,] 0.8000000
#> [356,] 0.8666667
#> [357,] 0.4666667
#> [358,] 0.9333333
#> [359,] 0.1333333
#> [360,] 0.6000000
#> [361,] 0.8000000
#> [362,] 0.8666667
#> [363,] 0.4666667
#> [364,] 0.7333333
#> [365,] 0.6666667
#> [366,] 0.7333333
#> [367,] 1.0000000
#> [368,] 0.4000000
#> [369,] 0.6666667
#> [370,] 0.6666667
#> [371,] 0.9333333
#> [372,] 0.7333333
#> [373,] 0.8666667
#> [374,] 0.6000000
#> [375,] 0.6000000
#> [376,] 0.6666667
#> [377,] 0.6666667
#> [378,] 0.6666667
#> [379,] 0.9333333
#> [380,] 0.8000000
#> [381,] 0.1333333
#> [382,] 0.4666667
#> [383,] 0.4666667
#> [384,] 0.4666667
#> [385,] 0.8000000
#> [386,] 0.5333333
#> [387,] 0.5333333
#> [388,] 0.4666667
#> [389,] 0.8000000
#> [390,] 0.6000000
#> [391,] 0.4000000
#> [392,] 0.4666667
#> [393,] 0.7333333
#> [394,] 0.4666667
#> [395,] 0.7333333
#> [396,] 0.5333333
#> [397,] 0.7333333
#> [398,] 0.8000000
#> [399,] 0.4000000
#> [400,] 0.3333333
#> [401,] 0.8666667
#> [402,] 0.4666667
#> [403,] 0.6666667
#> [404,] 0.8000000
#> [405,] 0.6666667
#> [406,] 0.6666667
#> [407,] 0.1333333
#> [408,] 0.7333333
#> [409,] 0.6666667
#> [410,] 0.8000000
#> [411,] 0.8000000
#> [412,] 0.6666667
#> [413,] 0.6666667
#> [414,] 0.6000000
#> [415,] 0.6666667
#> [416,] 1.0000000
#> [417,] 0.6000000
#> [418,] 0.8666667
#> [419,] 0.4000000
#> [420,] 0.6000000
#> [421,] 0.6000000
#> [422,] 0.4666667
#> [423,] 0.6000000
#> [424,] 0.8666667
#> [425,] 0.6000000
#> [426,] 0.6000000
#> [427,] 0.8666667
#> [428,] 0.3333333
#> [429,] 0.7333333
#> [430,] 0.7333333
#> [431,] 0.6666667
#> [432,] 0.8000000
#> [433,] 0.7333333
#> [434,] 0.8666667
#> [435,] 0.6666667
#> [436,] 0.9333333
#> [437,] 0.7333333
#> [438,] 0.6666667
#> [439,] 0.5333333
#> [440,] 0.7333333
#> [441,] 0.7333333
#> [442,] 0.6666667
#> [443,] 0.7333333
#> [444,] 0.7333333
#> [445,] 0.5333333
#> [446,] 0.4000000
#> [447,] 0.9333333
#> [448,] 0.8000000
#> [449,] 0.6666667
#> [450,] 0.8666667
#> [451,] 0.6666667
#> [452,] 0.8666667
#> [453,] 0.5333333
#> [454,] 0.8000000
#> [455,] 0.8666667
#> [456,] 0.2666667
#> [457,] 0.8666667
#> [458,] 0.7333333
#> [459,] 0.6666667
#> [460,] 0.4000000
#> [461,] 0.4666667
#> [462,] 0.5333333
#> [463,] 0.4000000
#> [464,] 0.2000000
#> [465,] 0.6666667
#> [466,] 0.8666667
#> [467,] 0.8000000
#> [468,] 0.8666667
#> [469,] 0.3333333
#> [470,] 0.4000000
#> [471,] 0.3333333
#> [472,] 0.4000000
#> [473,] 0.6666667
#> [474,] 0.4000000
#> [475,] 0.8000000
#> [476,] 0.6000000
#> [477,] 0.7333333
#> [478,] 0.9333333
#> [479,] 0.4666667
#> [480,] 0.8000000
#> [481,] 0.4666667
#> [482,] 0.8666667
#> [483,] 0.6000000
#> [484,] 0.7333333
#> [485,] 0.9333333
#> [486,] 0.8666667
#> [487,] 0.7333333
#> [488,] 0.7333333
#> [489,] 0.7333333
#> [490,] 0.6000000
#> [491,] 0.4000000
#> [492,] 0.8000000
#> [493,] 0.6000000
#> [494,] 0.6000000
#> [495,] 0.4666667
#> [496,] 0.4000000
#> [497,] 0.4666667
#> [498,] 0.4666667
#> [499,] 0.6666667
#> [500,] 0.4666667