Classical Test Theory

This function calculates the overall alpha and omega coefficients for the given data matrix. It also computes the alpha coefficient for each item, assuming that item is excluded.

Usage

CTT(U, na = NULL, Z = NULL, w = NULL)

Arguments

U

U is a data matrix of the type matrix or data.frame.

na

na argument specifies the numbers or characters to be treated as missing values.

Z

Z is a missing indicator matrix of the type matrix or data.frame

w

w is item weight vector

Value

Returns a list of class c("exametrika", "CTT") containing two data frames:

Reliability

A data frame with overall reliability coefficients (Alpha and Omega) calculated using different correlation matrices (Covariance, Phi, and Tetrachoric)

ReliabilityExcludingItem

A data frame showing alpha coefficients when each item is excluded, calculated using different correlation matrices

Examples

# \donttest{
# using sample dataset
CTT(J15S500)
# }