StudentAnalysis

The StudentAnalysis function returns descriptive statistics for each individual student. Specifically, it provides the number of responses, the number of correct answers, the passage rate, the standardized score, the percentile, and the stanine.

Usage

StudentAnalysis(U, na = NULL, Z = NULL, w = NULL)

Arguments

U

U is a data matrix of the type matrix or data.frame.

na

na argument specifies the numbers or characters to be treated as missing values.

ID: Student identifier

NR: Number of responses

NRS: Number-right score (total correct answers)

PR: Passage rate (proportion correct)

SS: Standardized score (z-score)

Percentile: Student's percentile rank

Stanine: Student's stanine score (1-9)

Z

Z is a missing indicator matrix of the type matrix or data.frame

w

w is item weight vector

Value

Returns a data frame containing the following columns for each student:

ID: Student identifier

NR: Number of responses

NRS: Number-right score (total correct answers)

PR: Passage rate (proportion correct)

SS: Standardized score (z-score)

Percentile: Student's percentile rank

Stanine: Student's stanine score (1-9)

Examples

# using sample dataset
StudentAnalysis(J15S500)
#>             ID NR NRS        PR         SS Percentile Stanine
#> 1   Student001 15   8 0.5333333 -0.6249599         32       4
#> 2   Student002 15  10 0.6666667  0.1261938         57       5
#> 3   Student003 15  10 0.6666667  0.1261938         57       5
#> 4   Student004 15  11 0.7333333  0.5017707         74       6
#> 5   Student005 15   6 0.4000000 -1.3761137         15       3
#> 6   Student006 15  12 0.8000000  0.8773476         87       7
#> 7   Student007 15   9 0.6000000 -0.2493831         42       5
#> 8   Student008 15  10 0.6666667  0.1261938         57       5
#> 9   Student009 15  14 0.9333333  1.6285014         99       9
#> 10  Student010 15   9 0.6000000 -0.2493831         42       5
#> 11  Student011 15   9 0.6000000 -0.2493831         42       5
#> 12  Student012 15   8 0.5333333 -0.6249599         32       4
#> 13  Student013 15  13 0.8666667  1.2529245         96       8
#> 14  Student014 15  12 0.8000000  0.8773476         87       7
#> 15  Student015 15   5 0.3333333 -1.7516906          8       2
#> 16  Student016 15   5 0.3333333 -1.7516906          8       2
#> 17  Student017 15  12 0.8000000  0.8773476         87       7
#> 18  Student018 15   7 0.4666667 -1.0005368         24       4
#> 19  Student019 15  12 0.8000000  0.8773476         87       7
#> 20  Student020 15   6 0.4000000 -1.3761137         15       3
#> 21  Student021 15  13 0.8666667  1.2529245         96       8
#> 22  Student022 15   8 0.5333333 -0.6249599         32       4
#> 23  Student023 15  10 0.6666667  0.1261938         57       5
#> 24  Student024 15  10 0.6666667  0.1261938         57       5
#> 25  Student025 15  10 0.6666667  0.1261938         57       5
#> 26  Student026 15   5 0.3333333 -1.7516906          8       2
#> 27  Student027 15   9 0.6000000 -0.2493831         42       5
#> 28  Student028 15  11 0.7333333  0.5017707         74       6
#> 29  Student029 15  11 0.7333333  0.5017707         74       6
#> 30  Student030 15   4 0.2666667 -2.1272675          3       1
#> 31  Student031 15   5 0.3333333 -1.7516906          8       2
#> 32  Student032 15  11 0.7333333  0.5017707         74       6
#> 33  Student033 15   5 0.3333333 -1.7516906          8       2
#> 34  Student034 15   5 0.3333333 -1.7516906          8       2
#> 35  Student035 15  15 1.0000000  2.0040783        100       9
#> 36  Student036 15   8 0.5333333 -0.6249599         32       4
#> 37  Student037 15  12 0.8000000  0.8773476         87       7
#> 38  Student038 15  11 0.7333333  0.5017707         74       6
#> 39  Student039 15   8 0.5333333 -0.6249599         32       4
#> 40  Student040 15  10 0.6666667  0.1261938         57       5
#> 41  Student041 15  11 0.7333333  0.5017707         74       6
#> 42  Student042 15  14 0.9333333  1.6285014         99       9
#> 43  Student043 15  10 0.6666667  0.1261938         57       5
#> 44  Student044 15   9 0.6000000 -0.2493831         42       5
#> 45  Student045 15  12 0.8000000  0.8773476         87       7
#> 46  Student046 15  11 0.7333333  0.5017707         74       6
#> 47  Student047 15  10 0.6666667  0.1261938         57       5
#> 48  Student048 15  12 0.8000000  0.8773476         87       7
#> 49  Student049 15  11 0.7333333  0.5017707         74       6
#> 50  Student050 15  13 0.8666667  1.2529245         96       8
#> 51  Student051 15  11 0.7333333  0.5017707         74       6
#> 52  Student052 15  12 0.8000000  0.8773476         87       7
#> 53  Student053 15  10 0.6666667  0.1261938         57       5
#> 54  Student054 15  11 0.7333333  0.5017707         74       6
#> 55  Student055 15  13 0.8666667  1.2529245         96       8
#> 56  Student056 15  12 0.8000000  0.8773476         87       7
#> 57  Student057 15   9 0.6000000 -0.2493831         42       5
#> 58  Student058 15  11 0.7333333  0.5017707         74       6
#> 59  Student059 15  14 0.9333333  1.6285014         99       9
#> 60  Student060 15  12 0.8000000  0.8773476         87       7
#> 61  Student061 15  14 0.9333333  1.6285014         99       9
#> 62  Student062 15  11 0.7333333  0.5017707         74       6
#> 63  Student063 15   9 0.6000000 -0.2493831         42       5
#> 64  Student064 15  11 0.7333333  0.5017707         74       6
#> 65  Student065 15  10 0.6666667  0.1261938         57       5
#> 66  Student066 15   5 0.3333333 -1.7516906          8       2
#> 67  Student067 15   7 0.4666667 -1.0005368         24       4
#> 68  Student068 15   6 0.4000000 -1.3761137         15       3
#> 69  Student069 15  12 0.8000000  0.8773476         87       7
#> 70  Student070 15   6 0.4000000 -1.3761137         15       3
#> 71  Student071 15   5 0.3333333 -1.7516906          8       2
#> 72  Student072 15  10 0.6666667  0.1261938         57       5
#> 73  Student073 15  10 0.6666667  0.1261938         57       5
#> 74  Student074 15  10 0.6666667  0.1261938         57       5
#> 75  Student075 15  12 0.8000000  0.8773476         87       7
#> 76  Student076 15   8 0.5333333 -0.6249599         32       4
#> 77  Student077 15  13 0.8666667  1.2529245         96       8
#> 78  Student078 15  12 0.8000000  0.8773476         87       7
#> 79  Student079 15  12 0.8000000  0.8773476         87       7
#> 80  Student080 15   4 0.2666667 -2.1272675          3       1
#> 81  Student081 15   9 0.6000000 -0.2493831         42       5
#> 82  Student082 15   5 0.3333333 -1.7516906          8       2
#> 83  Student083 15   8 0.5333333 -0.6249599         32       4
#> 84  Student084 15   8 0.5333333 -0.6249599         32       4
#> 85  Student085 15  12 0.8000000  0.8773476         87       7
#> 86  Student086 15   8 0.5333333 -0.6249599         32       4
#> 87  Student087 15  11 0.7333333  0.5017707         74       6
#> 88  Student088 15   8 0.5333333 -0.6249599         32       4
#> 89  Student089 15  12 0.8000000  0.8773476         87       7
#> 90  Student090 15  11 0.7333333  0.5017707         74       6
#> 91  Student091 15   7 0.4666667 -1.0005368         24       4
#> 92  Student092 15   6 0.4000000 -1.3761137         15       3
#> 93  Student093 15   8 0.5333333 -0.6249599         32       4
#> 94  Student094 15  12 0.8000000  0.8773476         87       7
#> 95  Student095 15   8 0.5333333 -0.6249599         32       4
#> 96  Student096 15  11 0.7333333  0.5017707         74       6
#> 97  Student097 15  10 0.6666667  0.1261938         57       5
#> 98  Student098 15   6 0.4000000 -1.3761137         15       3
#> 99  Student099 15  11 0.7333333  0.5017707         74       6
#> 100 Student100 15  12 0.8000000  0.8773476         87       7
#> 101 Student101 15  10 0.6666667  0.1261938         57       5
#> 102 Student102 15   6 0.4000000 -1.3761137         15       3
#> 103 Student103 15  13 0.8666667  1.2529245         96       8
#> 104 Student104 15   5 0.3333333 -1.7516906          8       2
#> 105 Student105 15  11 0.7333333  0.5017707         74       6
#> 106 Student106 15   7 0.4666667 -1.0005368         24       4
#> 107 Student107 15  11 0.7333333  0.5017707         74       6
#> 108 Student108 15  11 0.7333333  0.5017707         74       6
#> 109 Student109 15   6 0.4000000 -1.3761137         15       3
#> 110 Student110 15  12 0.8000000  0.8773476         87       7
#> 111 Student111 15  12 0.8000000  0.8773476         87       7
#> 112 Student112 15  11 0.7333333  0.5017707         74       6
#> 113 Student113 15   6 0.4000000 -1.3761137         15       3
#> 114 Student114 15   8 0.5333333 -0.6249599         32       4
#> 115 Student115 15   5 0.3333333 -1.7516906          8       2
#> 116 Student116 15  13 0.8666667  1.2529245         96       8
#> 117 Student117 15  11 0.7333333  0.5017707         74       6
#> 118 Student118 15   7 0.4666667 -1.0005368         24       4
#> 119 Student119 15   8 0.5333333 -0.6249599         32       4
#> 120 Student120 15   8 0.5333333 -0.6249599         32       4
#> 121 Student121 15  10 0.6666667  0.1261938         57       5
#> 122 Student122 15   8 0.5333333 -0.6249599         32       4
#> 123 Student123 15  12 0.8000000  0.8773476         87       7
#> 124 Student124 15  13 0.8666667  1.2529245         96       8
#> 125 Student125 15   9 0.6000000 -0.2493831         42       5
#> 126 Student126 15  13 0.8666667  1.2529245         96       8
#> 127 Student127 15   4 0.2666667 -2.1272675          3       1
#> 128 Student128 15  11 0.7333333  0.5017707         74       6
#> 129 Student129 15  12 0.8000000  0.8773476         87       7
#> 130 Student130 15  12 0.8000000  0.8773476         87       7
#> 131 Student131 15  10 0.6666667  0.1261938         57       5
#> 132 Student132 15  10 0.6666667  0.1261938         57       5
#> 133 Student133 15  11 0.7333333  0.5017707         74       6
#> 134 Student134 15   5 0.3333333 -1.7516906          8       2
#> 135 Student135 15   8 0.5333333 -0.6249599         32       4
#> 136 Student136 15   5 0.3333333 -1.7516906          8       2
#> 137 Student137 15  13 0.8666667  1.2529245         96       8
#> 138 Student138 15  11 0.7333333  0.5017707         74       6
#> 139 Student139 15  11 0.7333333  0.5017707         74       6
#> 140 Student140 15  10 0.6666667  0.1261938         57       5
#> 141 Student141 15  11 0.7333333  0.5017707         74       6
#> 142 Student142 15   5 0.3333333 -1.7516906          8       2
#> 143 Student143 15  13 0.8666667  1.2529245         96       8
#> 144 Student144 15   4 0.2666667 -2.1272675          3       1
#> 145 Student145 15  12 0.8000000  0.8773476         87       7
#> 146 Student146 15   7 0.4666667 -1.0005368         24       4
#> 147 Student147 15   5 0.3333333 -1.7516906          8       2
#> 148 Student148 15  12 0.8000000  0.8773476         87       7
#> 149 Student149 15   9 0.6000000 -0.2493831         42       5
#> 150 Student150 15  13 0.8666667  1.2529245         96       8
#> 151 Student151 15  13 0.8666667  1.2529245         96       8
#> 152 Student152 15  11 0.7333333  0.5017707         74       6
#> 153 Student153 15  14 0.9333333  1.6285014         99       9
#> 154 Student154 15   9 0.6000000 -0.2493831         42       5
#> 155 Student155 15   6 0.4000000 -1.3761137         15       3
#> 156 Student156 15  11 0.7333333  0.5017707         74       6
#> 157 Student157 15  10 0.6666667  0.1261938         57       5
#> 158 Student158 15  13 0.8666667  1.2529245         96       8
#> 159 Student159 15  13 0.8666667  1.2529245         96       8
#> 160 Student160 15  15 1.0000000  2.0040783        100       9
#> 161 Student161 15  11 0.7333333  0.5017707         74       6
#> 162 Student162 15   7 0.4666667 -1.0005368         24       4
#> 163 Student163 15   7 0.4666667 -1.0005368         24       4
#> 164 Student164 15   5 0.3333333 -1.7516906          8       2
#> 165 Student165 15  10 0.6666667  0.1261938         57       5
#> 166 Student166 15   5 0.3333333 -1.7516906          8       2
#> 167 Student167 15  10 0.6666667  0.1261938         57       5
#> 168 Student168 15   8 0.5333333 -0.6249599         32       4
#> 169 Student169 15  14 0.9333333  1.6285014         99       9
#> 170 Student170 15   6 0.4000000 -1.3761137         15       3
#> 171 Student171 15   6 0.4000000 -1.3761137         15       3
#> 172 Student172 15  13 0.8666667  1.2529245         96       8
#> 173 Student173 15  10 0.6666667  0.1261938         57       5
#> 174 Student174 15  12 0.8000000  0.8773476         87       7
#> 175 Student175 15   7 0.4666667 -1.0005368         24       4
#> 176 Student176 15  10 0.6666667  0.1261938         57       5
#> 177 Student177 15  11 0.7333333  0.5017707         74       6
#> 178 Student178 15   8 0.5333333 -0.6249599         32       4
#> 179 Student179 15  10 0.6666667  0.1261938         57       5
#> 180 Student180 15   8 0.5333333 -0.6249599         32       4
#> 181 Student181 15   7 0.4666667 -1.0005368         24       4
#> 182 Student182 15   6 0.4000000 -1.3761137         15       3
#> 183 Student183 15  11 0.7333333  0.5017707         74       6
#> 184 Student184 15   8 0.5333333 -0.6249599         32       4
#> 185 Student185 15  11 0.7333333  0.5017707         74       6
#> 186 Student186 15  11 0.7333333  0.5017707         74       6
#> 187 Student187 15   8 0.5333333 -0.6249599         32       4
#> 188 Student188 15  11 0.7333333  0.5017707         74       6
#> 189 Student189 15   9 0.6000000 -0.2493831         42       5
#> 190 Student190 15  10 0.6666667  0.1261938         57       5
#> 191 Student191 15  10 0.6666667  0.1261938         57       5
#> 192 Student192 15   7 0.4666667 -1.0005368         24       4
#> 193 Student193 15  10 0.6666667  0.1261938         57       5
#> 194 Student194 15   6 0.4000000 -1.3761137         15       3
#> 195 Student195 15   3 0.2000000 -2.5028444          1       1
#> 196 Student196 15  12 0.8000000  0.8773476         87       7
#> 197 Student197 15  11 0.7333333  0.5017707         74       6
#> 198 Student198 15  13 0.8666667  1.2529245         96       8
#> 199 Student199 15  10 0.6666667  0.1261938         57       5
#> 200 Student200 15  11 0.7333333  0.5017707         74       6
#> 201 Student201 15  12 0.8000000  0.8773476         87       7
#> 202 Student202 15  11 0.7333333  0.5017707         74       6
#> 203 Student203 15   5 0.3333333 -1.7516906          8       2
#> 204 Student204 15  12 0.8000000  0.8773476         87       7
#> 205 Student205 15  10 0.6666667  0.1261938         57       5
#> 206 Student206 15   8 0.5333333 -0.6249599         32       4
#> 207 Student207 15  12 0.8000000  0.8773476         87       7
#> 208 Student208 15  11 0.7333333  0.5017707         74       6
#> 209 Student209 15   9 0.6000000 -0.2493831         42       5
#> 210 Student210 15  11 0.7333333  0.5017707         74       6
#> 211 Student211 15  10 0.6666667  0.1261938         57       5
#> 212 Student212 15  14 0.9333333  1.6285014         99       9
#> 213 Student213 15  10 0.6666667  0.1261938         57       5
#> 214 Student214 15   8 0.5333333 -0.6249599         32       4
#> 215 Student215 15  10 0.6666667  0.1261938         57       5
#> 216 Student216 15   9 0.6000000 -0.2493831         42       5
#> 217 Student217 15  13 0.8666667  1.2529245         96       8
#> 218 Student218 15  12 0.8000000  0.8773476         87       7
#> 219 Student219 15   4 0.2666667 -2.1272675          3       1
#> 220 Student220 15   6 0.4000000 -1.3761137         15       3
#> 221 Student221 15   5 0.3333333 -1.7516906          8       2
#> 222 Student222 15  11 0.7333333  0.5017707         74       6
#> 223 Student223 15   9 0.6000000 -0.2493831         42       5
#> 224 Student224 15   9 0.6000000 -0.2493831         42       5
#> 225 Student225 15  10 0.6666667  0.1261938         57       5
#> 226 Student226 15  10 0.6666667  0.1261938         57       5
#> 227 Student227 15   9 0.6000000 -0.2493831         42       5
#> 228 Student228 15  11 0.7333333  0.5017707         74       6
#> 229 Student229 15   7 0.4666667 -1.0005368         24       4
#> 230 Student230 15  10 0.6666667  0.1261938         57       5
#> 231 Student231 15   7 0.4666667 -1.0005368         24       4
#> 232 Student232 15   4 0.2666667 -2.1272675          3       1
#> 233 Student233 15  12 0.8000000  0.8773476         87       7
#> 234 Student234 15   4 0.2666667 -2.1272675          3       1
#> 235 Student235 15   9 0.6000000 -0.2493831         42       5
#> 236 Student236 15   9 0.6000000 -0.2493831         42       5
#> 237 Student237 15   7 0.4666667 -1.0005368         24       4
#> 238 Student238 15  14 0.9333333  1.6285014         99       9
#> 239 Student239 15  12 0.8000000  0.8773476         87       7
#> 240 Student240 15  11 0.7333333  0.5017707         74       6
#> 241 Student241 15  12 0.8000000  0.8773476         87       7
#> 242 Student242 15  11 0.7333333  0.5017707         74       6
#> 243 Student243 15   9 0.6000000 -0.2493831         42       5
#> 244 Student244 15   8 0.5333333 -0.6249599         32       4
#> 245 Student245 15  11 0.7333333  0.5017707         74       6
#> 246 Student246 15   5 0.3333333 -1.7516906          8       2
#> 247 Student247 15  11 0.7333333  0.5017707         74       6
#> 248 Student248 15  15 1.0000000  2.0040783        100       9
#> 249 Student249 15   8 0.5333333 -0.6249599         32       4
#> 250 Student250 15  14 0.9333333  1.6285014         99       9
#> 251 Student251 15   9 0.6000000 -0.2493831         42       5
#> 252 Student252 15   7 0.4666667 -1.0005368         24       4
#> 253 Student253 15  11 0.7333333  0.5017707         74       6
#> 254 Student254 15  10 0.6666667  0.1261938         57       5
#> 255 Student255 15  11 0.7333333  0.5017707         74       6
#> 256 Student256 15  10 0.6666667  0.1261938         57       5
#> 257 Student257 15   8 0.5333333 -0.6249599         32       4
#> 258 Student258 15   7 0.4666667 -1.0005368         24       4
#> 259 Student259 15   7 0.4666667 -1.0005368         24       4
#> 260 Student260 15   9 0.6000000 -0.2493831         42       5
#> 261 Student261 15   7 0.4666667 -1.0005368         24       4
#> 262 Student262 15  13 0.8666667  1.2529245         96       8
#> 263 Student263 15  12 0.8000000  0.8773476         87       7
#> 264 Student264 15   8 0.5333333 -0.6249599         32       4
#> 265 Student265 15  12 0.8000000  0.8773476         87       7
#> 266 Student266 15  11 0.7333333  0.5017707         74       6
#> 267 Student267 15  11 0.7333333  0.5017707         74       6
#> 268 Student268 15   8 0.5333333 -0.6249599         32       4
#> 269 Student269 15  12 0.8000000  0.8773476         87       7
#> 270 Student270 15   8 0.5333333 -0.6249599         32       4
#> 271 Student271 15  12 0.8000000  0.8773476         87       7
#> 272 Student272 15   9 0.6000000 -0.2493831         42       5
#> 273 Student273 15   6 0.4000000 -1.3761137         15       3
#> 274 Student274 15   9 0.6000000 -0.2493831         42       5
#> 275 Student275 15  10 0.6666667  0.1261938         57       5
#> 276 Student276 15  11 0.7333333  0.5017707         74       6
#> 277 Student277 15   7 0.4666667 -1.0005368         24       4
#> 278 Student278 15   6 0.4000000 -1.3761137         15       3
#> 279 Student279 15  10 0.6666667  0.1261938         57       5
#> 280 Student280 15   7 0.4666667 -1.0005368         24       4
#> 281 Student281 15  10 0.6666667  0.1261938         57       5
#> 282 Student282 15  12 0.8000000  0.8773476         87       7
#> 283 Student283 15   7 0.4666667 -1.0005368         24       4
#> 284 Student284 15  12 0.8000000  0.8773476         87       7
#> 285 Student285 15   9 0.6000000 -0.2493831         42       5
#> 286 Student286 15   6 0.4000000 -1.3761137         15       3
#> 287 Student287 15  13 0.8666667  1.2529245         96       8
#> 288 Student288 15   9 0.6000000 -0.2493831         42       5
#> 289 Student289 15  13 0.8666667  1.2529245         96       8
#> 290 Student290 15  15 1.0000000  2.0040783        100       9
#> 291 Student291 15  10 0.6666667  0.1261938         57       5
#> 292 Student292 15  14 0.9333333  1.6285014         99       9
#> 293 Student293 15   4 0.2666667 -2.1272675          3       1
#> 294 Student294 15  12 0.8000000  0.8773476         87       7
#> 295 Student295 15   6 0.4000000 -1.3761137         15       3
#> 296 Student296 15   7 0.4666667 -1.0005368         24       4
#> 297 Student297 15  11 0.7333333  0.5017707         74       6
#> 298 Student298 15  12 0.8000000  0.8773476         87       7
#> 299 Student299 15   5 0.3333333 -1.7516906          8       2
#> 300 Student300 15  11 0.7333333  0.5017707         74       6
#> 301 Student301 15   7 0.4666667 -1.0005368         24       4
#> 302 Student302 15  13 0.8666667  1.2529245         96       8
#> 303 Student303 15   9 0.6000000 -0.2493831         42       5
#> 304 Student304 15   9 0.6000000 -0.2493831         42       5
#> 305 Student305 15  10 0.6666667  0.1261938         57       5
#> 306 Student306 15  12 0.8000000  0.8773476         87       7
#> 307 Student307 15   6 0.4000000 -1.3761137         15       3
#> 308 Student308 15  12 0.8000000  0.8773476         87       7
#> 309 Student309 15  11 0.7333333  0.5017707         74       6
#> 310 Student310 15   9 0.6000000 -0.2493831         42       5
#> 311 Student311 15  10 0.6666667  0.1261938         57       5
#> 312 Student312 15  11 0.7333333  0.5017707         74       6
#> 313 Student313 15  12 0.8000000  0.8773476         87       7
#> 314 Student314 15   7 0.4666667 -1.0005368         24       4
#> 315 Student315 15  11 0.7333333  0.5017707         74       6
#> 316 Student316 15  11 0.7333333  0.5017707         74       6
#> 317 Student317 15  13 0.8666667  1.2529245         96       8
#> 318 Student318 15   7 0.4666667 -1.0005368         24       4
#> 319 Student319 15   6 0.4000000 -1.3761137         15       3
#> 320 Student320 15   9 0.6000000 -0.2493831         42       5
#> 321 Student321 15   9 0.6000000 -0.2493831         42       5
#> 322 Student322 15  13 0.8666667  1.2529245         96       8
#> 323 Student323 15   9 0.6000000 -0.2493831         42       5
#> 324 Student324 15  13 0.8666667  1.2529245         96       8
#> 325 Student325 15  13 0.8666667  1.2529245         96       8
#> 326 Student326 15  13 0.8666667  1.2529245         96       8
#> 327 Student327 15   8 0.5333333 -0.6249599         32       4
#> 328 Student328 15  10 0.6666667  0.1261938         57       5
#> 329 Student329 15  11 0.7333333  0.5017707         74       6
#> 330 Student330 15  10 0.6666667  0.1261938         57       5
#> 331 Student331 15  12 0.8000000  0.8773476         87       7
#> 332 Student332 15   7 0.4666667 -1.0005368         24       4
#> 333 Student333 15   9 0.6000000 -0.2493831         42       5
#> 334 Student334 15  13 0.8666667  1.2529245         96       8
#> 335 Student335 15  11 0.7333333  0.5017707         74       6
#> 336 Student336 15  10 0.6666667  0.1261938         57       5
#> 337 Student337 15  10 0.6666667  0.1261938         57       5
#> 338 Student338 15  12 0.8000000  0.8773476         87       7
#> 339 Student339 15  13 0.8666667  1.2529245         96       8
#> 340 Student340 15   9 0.6000000 -0.2493831         42       5
#> 341 Student341 15   4 0.2666667 -2.1272675          3       1
#> 342 Student342 15  10 0.6666667  0.1261938         57       5
#> 343 Student343 15  10 0.6666667  0.1261938         57       5
#> 344 Student344 15  14 0.9333333  1.6285014         99       9
#> 345 Student345 15   7 0.4666667 -1.0005368         24       4
#> 346 Student346 15  11 0.7333333  0.5017707         74       6
#> 347 Student347 15  10 0.6666667  0.1261938         57       5
#> 348 Student348 15  12 0.8000000  0.8773476         87       7
#> 349 Student349 15  10 0.6666667  0.1261938         57       5
#> 350 Student350 15  12 0.8000000  0.8773476         87       7
#> 351 Student351 15  10 0.6666667  0.1261938         57       5
#> 352 Student352 15   8 0.5333333 -0.6249599         32       4
#> 353 Student353 15  11 0.7333333  0.5017707         74       6
#> 354 Student354 15   8 0.5333333 -0.6249599         32       4
#> 355 Student355 15  12 0.8000000  0.8773476         87       7
#> 356 Student356 15  13 0.8666667  1.2529245         96       8
#> 357 Student357 15   7 0.4666667 -1.0005368         24       4
#> 358 Student358 15  14 0.9333333  1.6285014         99       9
#> 359 Student359 15   2 0.1333333 -2.8784213          1       1
#> 360 Student360 15   9 0.6000000 -0.2493831         42       5
#> 361 Student361 15  12 0.8000000  0.8773476         87       7
#> 362 Student362 15  13 0.8666667  1.2529245         96       8
#> 363 Student363 15   7 0.4666667 -1.0005368         24       4
#> 364 Student364 15  11 0.7333333  0.5017707         74       6
#> 365 Student365 15  10 0.6666667  0.1261938         57       5
#> 366 Student366 15  11 0.7333333  0.5017707         74       6
#> 367 Student367 15  15 1.0000000  2.0040783        100       9
#> 368 Student368 15   6 0.4000000 -1.3761137         15       3
#> 369 Student369 15  10 0.6666667  0.1261938         57       5
#> 370 Student370 15  10 0.6666667  0.1261938         57       5
#> 371 Student371 15  14 0.9333333  1.6285014         99       9
#> 372 Student372 15  11 0.7333333  0.5017707         74       6
#> 373 Student373 15  13 0.8666667  1.2529245         96       8
#> 374 Student374 15   9 0.6000000 -0.2493831         42       5
#> 375 Student375 15   9 0.6000000 -0.2493831         42       5
#> 376 Student376 15  10 0.6666667  0.1261938         57       5
#> 377 Student377 15  10 0.6666667  0.1261938         57       5
#> 378 Student378 15  10 0.6666667  0.1261938         57       5
#> 379 Student379 15  14 0.9333333  1.6285014         99       9
#> 380 Student380 15  12 0.8000000  0.8773476         87       7
#> 381 Student381 15   2 0.1333333 -2.8784213          1       1
#> 382 Student382 15   7 0.4666667 -1.0005368         24       4
#> 383 Student383 15   7 0.4666667 -1.0005368         24       4
#> 384 Student384 15   7 0.4666667 -1.0005368         24       4
#> 385 Student385 15  12 0.8000000  0.8773476         87       7
#> 386 Student386 15   8 0.5333333 -0.6249599         32       4
#> 387 Student387 15   8 0.5333333 -0.6249599         32       4
#> 388 Student388 15   7 0.4666667 -1.0005368         24       4
#> 389 Student389 15  12 0.8000000  0.8773476         87       7
#> 390 Student390 15   9 0.6000000 -0.2493831         42       5
#> 391 Student391 15   6 0.4000000 -1.3761137         15       3
#> 392 Student392 15   7 0.4666667 -1.0005368         24       4
#> 393 Student393 15  11 0.7333333  0.5017707         74       6
#> 394 Student394 15   7 0.4666667 -1.0005368         24       4
#> 395 Student395 15  11 0.7333333  0.5017707         74       6
#> 396 Student396 15   8 0.5333333 -0.6249599         32       4
#> 397 Student397 15  11 0.7333333  0.5017707         74       6
#> 398 Student398 15  12 0.8000000  0.8773476         87       7
#> 399 Student399 15   6 0.4000000 -1.3761137         15       3
#> 400 Student400 15   5 0.3333333 -1.7516906          8       2
#> 401 Student401 15  13 0.8666667  1.2529245         96       8
#> 402 Student402 15   7 0.4666667 -1.0005368         24       4
#> 403 Student403 15  10 0.6666667  0.1261938         57       5
#> 404 Student404 15  12 0.8000000  0.8773476         87       7
#> 405 Student405 15  10 0.6666667  0.1261938         57       5
#> 406 Student406 15  10 0.6666667  0.1261938         57       5
#> 407 Student407 15   2 0.1333333 -2.8784213          1       1
#> 408 Student408 15  11 0.7333333  0.5017707         74       6
#> 409 Student409 15  10 0.6666667  0.1261938         57       5
#> 410 Student410 15  12 0.8000000  0.8773476         87       7
#> 411 Student411 15  12 0.8000000  0.8773476         87       7
#> 412 Student412 15  10 0.6666667  0.1261938         57       5
#> 413 Student413 15  10 0.6666667  0.1261938         57       5
#> 414 Student414 15   9 0.6000000 -0.2493831         42       5
#> 415 Student415 15  10 0.6666667  0.1261938         57       5
#> 416 Student416 15  15 1.0000000  2.0040783        100       9
#> 417 Student417 15   9 0.6000000 -0.2493831         42       5
#> 418 Student418 15  13 0.8666667  1.2529245         96       8
#> 419 Student419 15   6 0.4000000 -1.3761137         15       3
#> 420 Student420 15   9 0.6000000 -0.2493831         42       5
#> 421 Student421 15   9 0.6000000 -0.2493831         42       5
#> 422 Student422 15   7 0.4666667 -1.0005368         24       4
#> 423 Student423 15   9 0.6000000 -0.2493831         42       5
#> 424 Student424 15  13 0.8666667  1.2529245         96       8
#> 425 Student425 15   9 0.6000000 -0.2493831         42       5
#> 426 Student426 15   9 0.6000000 -0.2493831         42       5
#> 427 Student427 15  13 0.8666667  1.2529245         96       8
#> 428 Student428 15   5 0.3333333 -1.7516906          8       2
#> 429 Student429 15  11 0.7333333  0.5017707         74       6
#> 430 Student430 15  11 0.7333333  0.5017707         74       6
#> 431 Student431 15  10 0.6666667  0.1261938         57       5
#> 432 Student432 15  12 0.8000000  0.8773476         87       7
#> 433 Student433 15  11 0.7333333  0.5017707         74       6
#> 434 Student434 15  13 0.8666667  1.2529245         96       8
#> 435 Student435 15  10 0.6666667  0.1261938         57       5
#> 436 Student436 15  14 0.9333333  1.6285014         99       9
#> 437 Student437 15  11 0.7333333  0.5017707         74       6
#> 438 Student438 15  10 0.6666667  0.1261938         57       5
#> 439 Student439 15   8 0.5333333 -0.6249599         32       4
#> 440 Student440 15  11 0.7333333  0.5017707         74       6
#> 441 Student441 15  11 0.7333333  0.5017707         74       6
#> 442 Student442 15  10 0.6666667  0.1261938         57       5
#> 443 Student443 15  11 0.7333333  0.5017707         74       6
#> 444 Student444 15  11 0.7333333  0.5017707         74       6
#> 445 Student445 15   8 0.5333333 -0.6249599         32       4
#> 446 Student446 15   6 0.4000000 -1.3761137         15       3
#> 447 Student447 15  14 0.9333333  1.6285014         99       9
#> 448 Student448 15  12 0.8000000  0.8773476         87       7
#> 449 Student449 15  10 0.6666667  0.1261938         57       5
#> 450 Student450 15  13 0.8666667  1.2529245         96       8
#> 451 Student451 15  10 0.6666667  0.1261938         57       5
#> 452 Student452 15  13 0.8666667  1.2529245         96       8
#> 453 Student453 15   8 0.5333333 -0.6249599         32       4
#> 454 Student454 15  12 0.8000000  0.8773476         87       7
#> 455 Student455 15  13 0.8666667  1.2529245         96       8
#> 456 Student456 15   4 0.2666667 -2.1272675          3       1
#> 457 Student457 15  13 0.8666667  1.2529245         96       8
#> 458 Student458 15  11 0.7333333  0.5017707         74       6
#> 459 Student459 15  10 0.6666667  0.1261938         57       5
#> 460 Student460 15   6 0.4000000 -1.3761137         15       3
#> 461 Student461 15   7 0.4666667 -1.0005368         24       4
#> 462 Student462 15   8 0.5333333 -0.6249599         32       4
#> 463 Student463 15   6 0.4000000 -1.3761137         15       3
#> 464 Student464 15   3 0.2000000 -2.5028444          1       1
#> 465 Student465 15  10 0.6666667  0.1261938         57       5
#> 466 Student466 15  13 0.8666667  1.2529245         96       8
#> 467 Student467 15  12 0.8000000  0.8773476         87       7
#> 468 Student468 15  13 0.8666667  1.2529245         96       8
#> 469 Student469 15   5 0.3333333 -1.7516906          8       2
#> 470 Student470 15   6 0.4000000 -1.3761137         15       3
#> 471 Student471 15   5 0.3333333 -1.7516906          8       2
#> 472 Student472 15   6 0.4000000 -1.3761137         15       3
#> 473 Student473 15  10 0.6666667  0.1261938         57       5
#> 474 Student474 15   6 0.4000000 -1.3761137         15       3
#> 475 Student475 15  12 0.8000000  0.8773476         87       7
#> 476 Student476 15   9 0.6000000 -0.2493831         42       5
#> 477 Student477 15  11 0.7333333  0.5017707         74       6
#> 478 Student478 15  14 0.9333333  1.6285014         99       9
#> 479 Student479 15   7 0.4666667 -1.0005368         24       4
#> 480 Student480 15  12 0.8000000  0.8773476         87       7
#> 481 Student481 15   7 0.4666667 -1.0005368         24       4
#> 482 Student482 15  13 0.8666667  1.2529245         96       8
#> 483 Student483 15   9 0.6000000 -0.2493831         42       5
#> 484 Student484 15  11 0.7333333  0.5017707         74       6
#> 485 Student485 15  14 0.9333333  1.6285014         99       9
#> 486 Student486 15  13 0.8666667  1.2529245         96       8
#> 487 Student487 15  11 0.7333333  0.5017707         74       6
#> 488 Student488 15  11 0.7333333  0.5017707         74       6
#> 489 Student489 15  11 0.7333333  0.5017707         74       6
#> 490 Student490 15   9 0.6000000 -0.2493831         42       5
#> 491 Student491 15   6 0.4000000 -1.3761137         15       3
#> 492 Student492 15  12 0.8000000  0.8773476         87       7
#> 493 Student493 15   9 0.6000000 -0.2493831         42       5
#> 494 Student494 15   9 0.6000000 -0.2493831         42       5
#> 495 Student495 15   7 0.4666667 -1.0005368         24       4
#> 496 Student496 15   6 0.4000000 -1.3761137         15       3
#> 497 Student497 15   7 0.4666667 -1.0005368         24       4
#> 498 Student498 15   7 0.4666667 -1.0005368         24       4
#> 499 Student499 15  10 0.6666667  0.1261938         57       5
#> 500 Student500 15   7 0.4666667 -1.0005368         24       4